pyproject.toml
[project]
name = "rag-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "langchain",
    "langchain-text-splitters",
    "langchain-community",
    "langchain-openai",
    "langchain-chroma",
    "langgraph",
    "pypdf",
    "tiktoken",
]
